refactor: rename native modules
