+++ native/mod-manager/.gitignore
