+++ native/mod-manager/Cargo.toml
@@ -1,7 +1,7 @@
 
 [package]
 edition = "2021"
-name = "native-mod"
+name = "mod-manager"
 version = "0.1.0"
 
 [lib]
