+++ package.json
@@ -8,12 +8,12 @@
   "scripts": {
     "build:dll-injector": "cd native/dll-injector && pnpm install && pnpm run build",
     "build:extractor": "cd native/extractor && pnpm install && pnpm run build",
-    "build:native-mod": "cd native/native-mod && pnpm install && pnpm run build",
-    "build:native-mod-tools": "cd native/mod-tools && pnpm install && pnpm run build",
-    "build:native-util": "cd native/native-util && pnpm install && pnpm run build",
+    "build:mod-manager": "cd native/mod-manager && pnpm install && pnpm run build",
+    "build:mod-manager-tools": "cd native/mod-tools && pnpm install && pnpm run build",
+    "build:utils": "cd native/utils && pnpm install && pnpm run build",
     "build:native-static-glb": "cd native/static-glb && pnpm install && pnpm run build",
-    "build:native-fs": "cd native/native-fs && pnpm install && pnpm run build",
-    "build:native": "pnpm run build:dll-injector && pnpm run build:extractor && pnpm run build:native-mod && pnpm run build:native-mod-tools && pnpm run build:native-util && pnpm run build:native-static-glb && pnpm run build:native-fs",
+    "build:fs": "cd native/fs && pnpm install && pnpm run build",
+    "build:native": "pnpm run build:dll-injector && pnpm run build:extractor && pnpm run build:mod-manager && pnpm run build:mod-manager-tools && pnpm run build:utils && pnpm run build:native-static-glb && pnpm run build:fs",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
     "fmt": "oxfmt",
